poll-bot
Bot for taking a poll in a chat.
!poll [option1] [option2] ...
Start a poll with given options -- no limit to number of options
!vote [pollid] [option]
Each ship can only vote once
!endpoll [pollid]
Only the creator of a poll can end it
Polls get a numerical ID generated at creation and are stored in a json file in the directory that the bot is running in.